% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ncbi_genome.R
\name{download_ncbi_genome}
\alias{download_ncbi_genome}
\title{Download virus genome from NCBI API}
\usage{
download_ncbi_genome(tax_id, max_errors = 5L)
}
\arguments{
\item{tax_id}{The taxonomy ID of target virus.}

\item{max_errors}{Max try repeats.}
}
\value{
invisible(NULL)
}
\description{
Download virus genome from NCBI API
}
