% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize_fastp_json.R
\name{organize_fastp_json}
\alias{organize_fastp_json}
\title{organize_fastp_json}
\usage{
organize_fastp_json(json, col_split = ".", dump_cols_keyworkds = NULL)
}
\value{
The return value, if any, from executing the function.
}
\description{
A fct function
}
