% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize_fastqc_zip.R
\name{organize_fastqc_zip}
\alias{organize_fastqc_zip}
\title{organize_fastqc_zip}
\usage{
organize_fastqc_zip(sample, reads, type, path)
}
\value{
The return value, if any, from executing the function.
}
\description{
A fct function
}
