% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retry.R
\name{retry}
\alias{retry}
\title{Retry a function call}
\usage{
retry(
  expr,
  is_error = function(x) "try-error" \%in\% class(x),
  max_errors = 5L,
  sleep = 0
)
}
\arguments{
\item{expr}{The expression that may need retry.}

\item{is_error}{Function used to decide if the expr is errored.}

\item{max_errors}{Max try repeats.}

\item{sleep}{Time interval between retries.}
}
\value{
The value of expression if without error.
}
\description{
Retry a function call
}
