% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ncbi_genome.R
\name{download_ncbi_genome}
\alias{download_ncbi_genome}
\title{Download virus genome from NCBI API}
\usage{
download_ncbi_genome(tax_id, path)
}
\arguments{
\item{tax_id}{The taxonomy ID of target virus.}

\item{path}{Where to save the fasta file.}
}
\value{
invisible(NULL)
}
\description{
Download virus genome from NCBI API
}
